fall_2020/requirements.txt
Jinja2==2.11.2
MarkupSafe
Pillow==7.2.0
PyWavelets
lazy-object-proxy
Pygments==2.7.0
appnope==0.1.0
astroid==2.4.2
bleach==3.2.0
cycler==0.10.0
decorator==4.4.2
entrypoints==0.3
html5lib==1.1
ipykernel==5.3.4
ipython
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.5.2
jedi==0.17.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==6.0.3
nbformat==5.0.7
networkx==2.5
notebook==6.1.4
numpy==1.19.2
olefile==0.46
pandocfilters==1.4.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit
ptyprocess==0.6.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.7
scikit-image==0.17.2
scipy
simplegeneric==0.8.1
six==1.15.0
terminado==0.8.3
testpath==0.4.4
timeout-decorator==0.4.1
tornado==6.0.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1

